Updated Mac install instructions
Now that the Release feature is being used, curl can be used to fetch and install the binary.

The downside to this edit is that it embeds the current version number, which will need to be updated for each release. It appears that Github doesn't provide a URL for downloading the latest version.