Enable pause/interrupt + fix conditionals.
Added support for the "pause/interrupt" command, which simply requests that the debugger stop at the earliest opportunity.
Fixed an issue where conditional expressions might not be evalulated in the right context as well as now only accepting "true" explicitly, rather than accepting anything that is not "nil" or "false".